+++ src/main/java/com/coplaca/apirest/repository/ProductRepository.java
@@ -5,11 +5,13 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
-    List<Product> findByIsActiveTrueAndStockQuantityGreaterThan(Long stock);
+    List<Product> findByIsActiveTrueAndStockQuantityGreaterThan(BigDecimal stock);
     List<Product> findByCategoryId(Long categoryId);
     List<Product> findByIsActiveTrue();
