+++ src/main/java/com/coplaca/apirest/service/ProductService.java
@@ -3,6 +3,7 @@ package com.coplaca.apirest.service;
 import com.coplaca.apirest.dto.ProductDTO;
 import com.coplaca.apirest.entity.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductService {
@@ -13,5 +14,5 @@ public interface ProductService {
     Product createProduct(Product product);
     Product updateProduct(Long id, Product productDetails);
     void disableProduct(Long id);
-    ProductDTO adjustStock(Long id, Long quantityChange);
+    ProductDTO adjustStock(Long id, BigDecimal quantityChange);
 }
